# Speedup matmul when the matrices are vectors