--- a/src/kernels/backend_webgl.ts
+++ b/src/kernels/backend_webgl.ts
@@ -600,8 +600,19 @@ export class MathBackendWebGL implements KernelBackend {
     const outerShapeA = transposeA ? a.shape[2] : a.shape[1];
     const outerShapeB = transposeB ? b.shape[1] : b.shape[2];
 
+    const [batch, firstDim, ] = a.shape;
+    const [, , secondDim] = b.shape;
+
+    if (firstDim === 1 && secondDim === 1) {
+      const a2D = a.as2D(batch, -1);
+      const b2D = b.as2D(batch, -1);
+      // Since the matrices are vectors, it is faster to call mul().sum()
+      // because sum() is O(sqrt(N)) due to divide-and-conquer.
+      return this.multiply(a2D, b2D).sum(1 /* axis */).as3D(batch, 1, 1);
+    }
+
     // TODO(https://github.com/tensorflow/tfjs/issues/693): Support 3D tensors
-    if (a.shape[0] === 1 && b.shape[0] === 1) {
+    if (batch === 1) {
       const aSqueezed = a.as2D(a.shape[1], a.shape[2]);
       const bSqueezed = b.as2D(b.shape[1], b.shape[2]);