@dsmilkov 2018-11-05
Hold on for a bit, I'm going to generalize this further for any matrix-vector, vector-matrix product based on the excellent observation of @DirkToewe in https://github.com/tensorflow/tfjs/issues/881#issuecomment-436045897